Add assertions that the file exists and is a file before asserting the json content
also split extensions for `Path`s/`File`s into separate packages
