+++ kotest-assertions/kotest-assertions-json/src/jvmTest/kotlin/com/sksamuel/kotest/tests/json/JsonFileAssertionsTest.kt
@@ -1,5 +1,9 @@
 package com.sksamuel.kotest.tests.json
 
+import io.kotest.assertions.json.file.shouldEqualJson
+import io.kotest.assertions.json.file.shouldNotEqualJson
+import io.kotest.assertions.json.paths.shouldEqualJson
+import io.kotest.assertions.json.paths.shouldNotEqualJson
 import io.kotest.assertions.json.shouldEqualJson
 import io.kotest.assertions.json.shouldNotEqualJson
 import io.kotest.assertions.throwables.shouldThrow
